samples/pnp_simple_thermostat/requirements.txt
mj-azure-iot-pnp-device
varname
